Endpoint authentication

+++ src/main/java/com/project/AzubiHub/controller/UserController.java
@@ -1,19 +1,16 @@
 package com.project.AzubiHub.controller;
 
-import com.project.AzubiHub.dto.AuthResponse;
-import com.project.AzubiHub.dto.UpdateUserRequestDto;
+import com.project.AzubiHub.config.AuthUser;
+import com.project.AzubiHub.dto.*;
 import com.project.AzubiHub.enitty.LoginUserRequest;
 import com.project.AzubiHub.enitty.RegisterUserRequest;
-import com.project.AzubiHub.dto.RegisterUserRequestDto;
 import com.project.AzubiHub.enitty.UpdateUserRequest;
 import com.project.AzubiHub.enitty.User;
-import com.project.AzubiHub.dto.UserDto;
 import com.project.AzubiHub.mapper.UserMapper;
 import com.project.AzubiHub.service.UserService;
 import com.project.AzubiHub.util.JwtService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -24,19 +21,20 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
 import java.util.Optional;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/auth")
 @Validated
-@Slf4j
 public class UserController {
 
     private final UserService userService;
     private final UserMapper userMapper;
     private final JwtService jwtService;
     private final AuthenticationManager authenticationManager;
+    private final AuthUser authUser;
 
     @PostMapping("/register")
     public ResponseEntity<UserDto> createNewUser(
@@ -68,6 +66,7 @@ public class UserController {
     @PutMapping("/{Id}/update")
     public ResponseEntity<UserDto> loginUser(
             @PathVariable Long Id,  @Valid @RequestBody UpdateUserRequestDto updateUserRequestDto ) {
+        authUser.validateUserAccess(Id);
         UpdateUserRequest updateUserRequest = userMapper.toUpdate(updateUserRequestDto);
         User user = userService.updateUser(Id, updateUserRequest);
         UserDto userUpdated = userMapper.toDto(user);
@@ -77,6 +76,7 @@ public class UserController {
     @GetMapping("/{Id}/get")
     public ResponseEntity<UserDto> getUserById(
             @PathVariable Long Id ) {
+        authUser.validateUserAccess(Id);
         Optional<User> user = userService.findUserById(Id);
         if (user.isPresent()) {
             return ResponseEntity.ok(userMapper.toDto(user.get()));
@@ -87,10 +87,21 @@ public class UserController {
     @DeleteMapping("/{Id}/delete")
     public ResponseEntity<Void> deleteUserById(
             @PathVariable Long Id ) {
+        authUser.validateUserAccess(Id);
         userService.deleteUserById(Id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 
     }
 
+    @GetMapping("/all-users")
+    public ResponseEntity<List<UserDtoSummary>> findAllUser(){
+        authUser.getAuthenticatedUserOrThrow();
+        List<User> allUser = userService.getAllUser();
+        List<UserDtoSummary> collectAllUsers = allUser.stream()
+                .map(userMapper::toSummary)
+                .toList();
+
+        return ResponseEntity.status(HttpStatus.OK).body(collectAllUsers);
+    }
 
 }
